refactor(mobile): extract SilentErrorBoundary to shared component
Move the error boundary out of WalletConnectBadge into
src/components/ErrorBoundary/ so it can be reused by any
non-critical UI that should render null on error.
